valign the whole container div as is the original <pre>
For some css reasons I don't understand, the margin-top of the first child
of the <dd> element (when code is downloadable) has no effect but
the margin-top of the inner elements is respected. This leads to bug #6.
The solution is to use the <pre> margin-top for the container and to
set the margin-top of the pre elem to 0.